requirements.txt
requests
google-cloud-storage
google-cloud-bigquery
pandas
google-cloud-secretmanager